Adding confirmation dialog for deleting data

+++ app/src/main/java/nadirbasalamah/android/com/numoapps/admin/ui/nutritionists/NutritionistDetailActivity.kt
@@ -1,5 +1,6 @@
 package nadirbasalamah.android.com.numoapps.admin.ui.nutritionists
 
+import android.app.AlertDialog
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
@@ -37,13 +38,26 @@ class NutritionistDetailActivity : AppCompatActivity() {
         }
 
         btn_delete_nutritionist.setOnClickListener {
-            adminViewModel = ViewModelProvider(this, ViewModelProvider.NewInstanceFactory()).get(AdminViewModel::class.java)
-            adminViewModel.setContext(applicationContext)
-            adminViewModel.deleteNutritionist(nutritionist.id)?.observe(this, Observer {result ->
-                if(result?.status == true) {
-                    finish()
+            val builder = AlertDialog.Builder(this)
+            builder.setMessage(R.string.delete_confirmation)
+                .setPositiveButton(R.string.confirmation_yes
+                ) { dialog, id ->
+                    adminViewModel = ViewModelProvider(this, ViewModelProvider.NewInstanceFactory()).get(AdminViewModel::class.java)
+                    adminViewModel.setContext(applicationContext)
+                    adminViewModel.deleteNutritionist(nutritionist.id)?.observe(this, Observer {result ->
+                        if(result?.status == true) {
+                            finish()
+                        }
+                    })
+                    dialog.dismiss()
                 }
-            })
+                .setNegativeButton(R.string.confirmation_no
+                ) { dialog, id ->
+                    dialog.cancel()
+                }
+            // Create the AlertDialog object and return it
+            builder.create()
+            builder.show()
         }
     }
 }
